Utilizando JPA-Test para testar as operações de salvar e atualizar Autor

+++ src/main/java/io/github/arthur32p/libraryapi/model/Livro.java
@@ -2,8 +2,8 @@ package io.github.arthur32p.libraryapi.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -31,7 +31,7 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale = 2)
-    private Double preco;
+    private BigDecimal preco;
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
